BACKEND: Modularización del Sistema Selnic, antes llamado DMS, se completo el module document-mamagement, corrigiendo errores al ejecutar la APP y corrigiendo problema de dependencias
