+++ backend/src/modules/document-management/models/Tramite.model.js
@@ -1,5 +1,6 @@
 import { sequelize } from "../../../config/db.config.js";
-import { config } from "../config/parametros.config.js";
+import { config } from "../../../config/config.js";
+
 import { TramiteHistorialEstado } from "./TramiteHistorialEstado.model.js";
 import { TramiteAsignacion } from "./TramiteAsignacion.model.js";
 import { TramiteArchivo } from "./TramiteArchivo.model.js";
