+++ frontend/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@tanstack/react-table": "^8.20.6",
     "axios": "^1.7.8",
+    "exceljs": "^4.4.0",
     "file-saver": "^2.0.5",
     "jspdf": "^3.0.0",
     "jspdf-autotable": "^5.0.2",
@@ -20,8 +21,7 @@
     "react-dom": "^18.3.1",
     "react-icons": "^5.5.0",
     "react-router-dom": "^7.0.1",
-    "react-select": "^5.10.1",
-    "xlsx": "^0.18.5"
+    "react-select": "^5.10.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.0",
